+++ src/kde/kde5/RpOverlayIconPlugin.cpp
@@ -27,12 +27,10 @@
 
 // librpbase
 #include "librpbase/RomData.hpp"
+#include "librpbase/file/FileSystem.hpp"
 #include "librpbase/file/RpFile.hpp"
 #include "librpbase/config/Config.hpp"
-using LibRpBase::RomData;
-using LibRpBase::IRpFile;
-using LibRpBase::RpFile;
-using LibRpBase::Config;
+using namespace LibRpBase;
 
 // libromdata
 #include "libromdata/RomDataFactory.hpp"
@@ -116,9 +114,17 @@ QStringList RpOverlayIconPlugin::getOverlays(const QUrl &item)
 	}
 
 	// Single file, and it's local.
+	const QByteArray u8filename = filename.toUtf8();
+
+	// Check for "bad" file systems.
+	if (FileSystem::isOnBadFS(u8filename.constData(), config->enableThumbnailOnNetworkFS())) {
+		// This file is on a "bad" file system.
+		return sl;
+	}
+
 	// TODO: RpQFile wrapper.
 	// For now, using RpFile, which is an stdio wrapper.
-	RpFile *const file = new RpFile(Q2U8(filename), RpFile::FM_OPEN_READ_GZ);
+	RpFile *const file = new RpFile(u8filename.constData(), RpFile::FM_OPEN_READ_GZ);
 	if (!file->isOpen()) {
 		// Could not open the file.
 		file->unref();